items type fix after migration update

+++ pkg/models/models.go
@@ -11,12 +11,12 @@ type Items struct {
 	ID        int     `gorm:"primary_key" json:"id"`
 	Name      string  `json:"name"`
 	Type      string  `json:"type"`
-	Protein   int     `json:"protein"`
-	Carbs     int     `json:"carbs"`
-	Fats      int     `json:"fats"`
-	Calories  int     `json:"calories"`
+	Protein   float32 `json:"protein"`
+	Carbs     float32 `json:"carbs"`
+	Fats      float32 `json:"fats"`
+	Calories  float32 `json:"calories"`
 	CalPerPro float32 `json:"calperpro"`
-	Sodium    int     `json:"sodium"`
+	Sodium    float32 `json:"sodium"`
 	RID       int     `sql:"type:integer REFERENCES restaurants(id) ON DELETE CASCADE ON UPDATE CASCADE" json:"r_id"`
 }
